i58: Add Flippable HOC
