+++ src/library/components/OnboardingButtons.js
@@ -1,40 +1,26 @@
 import React from 'react';
-import { View } from 'react-native';
-import Animated from 'react-native-reanimated';
 import { connect } from 'react-redux';
 import { goToUser } from 'rdx/locationState';
 import { getLocation } from 'library/apis/geolocation';
-import { withWait } from 'library/components/hoc';
+import { Flippable } from 'library/components/hoc';
 import FAB from '../components/FAB';
 import TwitterAuth from '../components/TwitterAuth';
-import { useAnimation } from 'library/hooks';
 import { STRINGS, normalizeStyles } from 'res';
 
 const OnboardingButtons = ({ authLoading, setAuthLoading, storageLoaded, userVisible, goToUser }) => {
-  const [frontAnimation, backAnimation] = useAnimation('flip', userVisible);
-  const ViewWithWait = withWait(View);
-  const touchableProps = {
-    pointerEvents: userVisible ? 'none' : 'auto',
-  };
+  const showButtons = storageLoaded && !authLoading && userVisible !== null;
 
   return (
-    <ViewWithWait
-      waitFor={storageLoaded && !authLoading && userVisible !== null}
-      style={styles.flipContainer}
-      pointerEvents='box-none'
-    >
-      <Animated.View style={backAnimation} >
+    showButtons &&
+      <Flippable trigger={userVisible} containerStyle={styles.flipContainer} >
         <TwitterAuth onLoading={(state) => setAuthLoading(state)} />
-      </Animated.View>
-      <Animated.View style={frontAnimation} {...touchableProps} >
         <FAB
           label={STRINGS.button.goToLocation}
           theme='green'
           icon='crosshairs'
           onPress={() => getLocation(goToUser)}
         />
-      </Animated.View>
-    </ViewWithWait>
+      </Flippable>
   );
 };
