+++ src/library/components/hoc/index.js
@@ -1,2 +1,2 @@
-export { default as withWait } from './withWait';
 export { default as Pressable } from './Pressable';
+export { default as Flippable } from './Flippable';
